+++ packages/core/src/lib/components/T/utils/useCamera.svelte.ts
@@ -37,28 +37,25 @@ export const updateCamera = (
 }
 
 export const useCamera = (
-  getCamera: () => PerspectiveCamera | OrthographicCamera,
-  getManual: () => boolean,
-  getMakeDefault: () => boolean,
+  camera: () => PerspectiveCamera | OrthographicCamera,
+  manual: () => boolean,
+  makeDefault: () => boolean,
   props: () => Record<string, unknown>
 ) => {
   const { camera: defaultCamera, manual: defaultManual, makeDefaultCameras } = useDefaultCamera()
   const { invalidate } = useScheduler()
   const { size } = useDOM()
 
-  const camera = $derived(getCamera())
-  const manual = $derived(getManual())
-
   $effect.pre(() => {
-    if (!getMakeDefault()) {
+    if (!makeDefault()) {
       return
     }
 
-    const currentCamera = camera
+    const currentCamera = camera()
 
     makeDefaultCameras.add(currentCamera)
     defaultCamera.set(currentCamera)
-    defaultManual.set(manual)
+    defaultManual.set(manual())
     invalidate()
 
     return () => {
@@ -75,13 +72,15 @@ export const useCamera = (
   })
 
   $effect.pre(() => {
-    if (manual) {
+    if (manual()) {
       return
     }
 
+    const currentCamera = camera()
+
     for (const key in props()) {
       if (updateProjectionMatrixKeys.has(key)) {
-        camera.updateProjectionMatrix()
+        currentCamera.updateProjectionMatrix()
         invalidate()
         break
       }
@@ -89,10 +88,10 @@ export const useCamera = (
   })
 
   $effect.pre(() => {
-    if (getManual()) {
+    if (manual()) {
       return
     }
 
-    updateCamera(camera, size.current.width, size.current.height)
+    updateCamera(camera(), size.current.width, size.current.height)
   })
 }